### 2020-12-05 #2888 @whedon
[PRE REVIEW]: Tyssue, en Epithelium simulation library
**Submitting author:** @glyg (<a href="http://orcid.org/0000-0002-5444-5246">Guillaume Gay</a>)
**Repository:** <a href="https://github.com/damcb/tyssue" target ="_blank">https://github.com/damcb/tyssue</a>
**Version:** v0.8.0
