### 2020-12-05 2a691d5
Update paper.bib
Corrects DOIs according to @wehdon
