+++ paper.bib
@@ -2,6 +2,7 @@
 author = {Martin, Emmanuel and Theis, Sophie and Monier, Bruno and Gay, Guillaume and Suzanne, Magali},
 title = {{Mechanical control of morphogenesis robustness}},
 url = {https://www.biorxiv.org/content/10.1101/2020.01.06.896266v1).},
+doi = {10.1101/2020.01.06.896266}
 }
 
 @article{Gracia:2019,
@@ -206,7 +207,7 @@ volume = "81",
 pages = "191 - 248",
 year = "1983",
 issn = "0074-7696",
-doi = "https://doi.org/10.1016/S0074-7696(08)62339-6",
+doi = {10.1016/S0074-7696(08)62339-6},
 url = "http://www.sciencedirect.com/science/article/pii/S0074769608623396",
 author = "Hisao Honda",
 }
@@ -219,7 +220,7 @@ number = "3",
 pages = "523 - 543",
 year = "1978",
 issn = "0022-5193",
-doi = "https://doi.org/10.1016/0022-5193(78)90315-6",
+doi = {10.1016/0022-5193(78)90315-6},
 url = "http://www.sciencedirect.com/science/article/pii/0022519378903156",
 author = "Hisao Honda",
 }
