### 2020-12-05 759cbf0
Update paper.md (#214)
* adds figure details, adds 3D in figure
