+++ paper.md
@@ -30,26 +30,20 @@ bibliography: paper.bib
 The `tyssue` Python library seeks to provide a unified interface to implement bio-mechanical models of living tissues. It's main focus is on vertex based epithelium models. `tyssue` allows to model the mechanical behavior of 2D, apical 3D or full 3D epihelia based on the numerical resolution of the equations of motion for the mesh vertices. Biological process are modeled through changes in the topological and dynamical properties of the mesh. `tyssue` is a modular library. Starting with the same tissue geometry, the choice of constraints, energy potential terms and parameters increases the possibility to answer different biological questions and easily explore mechanical hypotheses.
 </div>
 
-
-<center>
-![figure1](doc/illus/figure1.jpg  "figure1")
-</center>
-
-
 # Statement of Need
 <div align="justify">
 
 Tissue remodelling is a complex process integrating a large number of input such as gene expression pattern, cell adherent properties, cell mechanics. It can be difficult to manipulate specific aspects genetically. It can even be hard to simply capture, when the process takes only few minutes. Furthermore, morphogenesis is inherently a mechanical process. To execute complex morphogenetic movements, epithelia are driven by in-plane forces, like constriction of apical cell surface [@Heer:2017], and/or out-plane forces, such as the apico-basal cable in apoptotic cell [@Monier:2015, @Gracia:2019] or lateral tension [@Sherrard:2010, @Sui:2018]. Modeling those process help us understand how tissue acquires their shape, in complement of the experimental systems, and beyond their limitations. Several vertex models have been developed in the past few years to describe the physics of epithelia (for a review, see [@Alt:2017]), and common features can be identified.
 
-The `tyssue` library defines tissue as meshes. A vertex model defines a tissue as an assembly of vertices and edges, which can form polygonal face (in 2D) or polyhedron (in 3D). For now, we assume that cell junction are straight lines. In `tyssue`, each edge is split, so that every face is limited by oriented "half-edges" (**figure2 A**), in a structure identical to the [Linear Cell Complex](https://doc.cgal.org/latest/Linear_cell_complex/index.html) in the CGAL library. The library implements concepts and mechanisms common to all vertex models, for both topological and mechanical aspects.
+The `tyssue` library defines tissue as meshes. A vertex model defines a tissue as an assembly of vertices and edges, which can form polygonal face (in 2D) or polyhedron (in 3D)  (see figure 1 B),. For now, we assume that cell junction are straight lines. In `tyssue`, each edge is split, so that every face is limited by oriented "half-edges" (see figure 1 A), in a structure identical to the [Linear Cell Complex](https://doc.cgal.org/latest/Linear_cell_complex/index.html) in the CGAL library. The library implements concepts and mechanisms common to all vertex models, for both topological and mechanical aspects.
 
 ### Topology
 
 Common cellular process are implemented in our library such as cell elimination, division or rearangements. We implemented those processes based on previous works.
 
-Cell division is modeled as the splitting of a cell by a straight line (or plane in 3D) [@Brodland:2002], the angle and position of the division plane can be decided.
+Cell division is modeled as the splitting of a cell by a straight line (or plane in 3D) [@Brodland:2002], the angle and position of the division plane can be decided (see Figure 1 C, right pannel).
 
-Changes in cell neighbours - also called rearangements - happen when the size of the boundary between two neighboring cells passes below a certain threshold length in 2D (type 1 transition), or area in 3D (I-H or H-I transition)  [@Okuda:2015]. In that case, the linked vertices fuse and are separated again, witch can lead to a change in the local topology.
[…]
