### 2020-12-07 #2888 @pdebuyl
@smsaladi can you review the article "Tyssue, en Epithelium simulation library" for JOSS?
