### 2020-12-10 #2888 @pdebuyl
@shriyaamittal can you review the article "Tyssue, en Epithelium simulation library" for the Journal of Open Source Software?
