### 2020-12-11 #2888 @pdebuyl
@Paijmans can you review the article "Tyssue, en Epithelium simulation library" for the Journal of Open Source Software?
