### 2020-12-15 #2888 @pdebuyl
@salbreux can you review the article "Tyssue, en Epithelium simulation library" for the Journal of Open Source Software? (please check our COI policy https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy ).
