### 2021-01-07 8ed8712
adds cgal and gmp to environment.yml
