+++ environment.yml
@@ -5,6 +5,8 @@ channels:
 dependencies:
 - boost-cpp
 - mpfr
+- gmp
+- cgal
 - python>=3.6
 - numpy
 - scipy
