### 2021-01-07 e19a8f0
fix write permission issue in config
