+++ tyssue/config/json_parser.py
@@ -11,7 +11,7 @@ logger = logging.getLogger(__name__)
 
 def load_spec(fname):
 
-    with open(fname, "r+") as config_file:
+    with open(fname, "r") as config_file:
         spec = json.load(config_file)
     return spec
 
