### 2021-01-19 #2973 @whedon
[REVIEW]: Tyssue, en Epithelium simulation library
**Submitting author:** @glyg (<a href="http://orcid.org/0000-0002-5444-5246">Guillaume Gay</a>)
**Repository:** <a href="https://github.com/damcb/tyssue" target ="_blank">https://github.com/damcb/tyssue</a>
**Version:** 0.9.0
