### 2021-01-27 b73428e
fix link to bibfile, closes #218
