+++ environment.yml
@@ -3,6 +3,7 @@ channels:
 - conda-forge
 
 dependencies:
+- pip
 - boost-cpp
 - mpfr
 - gmp
