### 2021-01-28 63562a0
bug fix (opening a config file in r+ mode) (#217)
* set release to true

* adds cgal and gmp to environment.yml

* fix write permission issue in config

* updates pybind11
