+++ .travis.yml
@@ -4,7 +4,7 @@ before_install:
   - sudo apt-get update
   - sudo apt-get install --yes build-essential
   - sudo apt-get install --yes git
-  - sudo apt-get install libcgal-dev libcgal-qt5-dev
+  - sudo apt-get install libmpfr-dev libgmp-dev libboost-all-dev
   - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
   - bash miniconda.sh -b -p $HOME/miniconda
   - export PATH="$HOME/miniconda/bin:$PATH"
@@ -16,7 +16,7 @@ before_install:
   - source activate tyssue
   - conda install pytest pytest-cov
   - pip install coveralls nbval
-  
+
 install:
   - python setup.py install
 
