+++ environment.yml
@@ -3,8 +3,11 @@ channels:
 - conda-forge
 
 dependencies:
+- pip
 - boost-cpp
 - mpfr
+- gmp
+- cgal
 - python>=3.6
 - numpy
 - scipy
