### 2021-02-02 #2973 @SergeDmi
# Review
## Software structure and installation
I have actually little to say about the software. It is open source, well documented, easy to install with conda (although conda may have its pitfalls, it worked at first try for me).

## Scholar effort
I think tyssue is a fantastic effort that really fills a need for well defined, reproducible physical models in the field, while allowing a lot of freedom for the user. It is clearly a fully-featured sofware.

## Software usage
Tyssue is an API, which allows a broad use, at the expense of simplicity for the non-programmer. My main concern is the lack of gradual difficulty in the tutorials : reproducing research results is great, but gently steering the user might be missing.
While example 5 is simple as claimed, example 2 is rather complex. To engage scientists to use the software, step-by-step tutorial, or a better gradation of difficulty in the notebook examples, seem required to smoothen the steep learning curve.

## Article
### Statement of need
The statement of need does a good job of explaining the context, but lacks in describing the actual need for the software. A mention of existing software is also lacking. Instead, this paragraph gives a short description of Tyssue.

### Software description
I think the article fails to convey the organization of Tyssue, that "Separates structure, geometry and models" (better explained in the readme). As a consequence, it is not clear that the models (referenced in the paragraph *mechanics*) can be chosen at will between Euler, Gradient descent, Quasi-static, and possibly other and user defined ; the same applies to *topology*.

# Recommendation
The software and article definitely should be published in JOSS. Usage examples should be streamlined. Article should highlight the need for this software as well as software organisation and capabilities. A link to the examples and doc should be provided.
