### 2021-02-02 #2973 @glyg
Hi @pdebuyl and @SergeDmi

Thank you a lot for your review Serge. @sophietheis and I are giving a workshop on tyssue next week, which will allow to test in the field a better tutorial.

If this is ok with you, we will answer Serge's remarks only after that (within the next 2 weeks).
