### 2021-02-22 #2973 @glyg
Hi @pdebuyl - The update to the doc is going on under this issue: https://github.com/DamCB/tyssue-demo/issues/4 - we are going back through the notebooks and trying to follow @SergeDmi 's advice to be more progressive. Also some newest notions as the `EventManager` haven't been documented properly yet.

As for the paper itself, we thought it might be better to have the comments from the second reviewer before we undertake the revisions.
