+++ tyssue/draw/plt_draw.py
@@ -12,7 +12,7 @@ import pathlib
 import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
-
+import matplotlib as mpl
 
 from matplotlib import cm
 from matplotlib.path import Path
@@ -163,28 +163,42 @@ def sheet_view_GC_colorbar(sheet, coords=COORDS, ax1=None, ax2=None, **draw_spec
 
     The default sheet_spec specification is:
 
-    {'edge': {
-      'visible': True,
-      'width': 0.5,
-      'head_width': 0.2, # arrow head width for the edges
-      'length_includes_head': True, # see matplotlib Arrow artist doc
-      'shape': 'right',
-      'color': '#2b5d0a', # can be an array
-      'alpha': 0.8,
-      'zorder': 1,
-      'colormap': 'viridis'},
-     'vert': {
-      'visible': True,
-      's': 100,
-      'color': '#000a4b',
-      'alpha': 0.3,
-      'zorder': 2},
-     'face': {
-      'visible': False,
-      'color': '#8aa678',
-      'alpha': 1.0,
-      'zorder': -1}
-      }
+    {"edge": {
+	"visible": true,
+    "width": 0.5,
+    "head_width": 0.0,
+    "length_includes_head": true,
+    "shape": "right",
+    "color": "#2b5d0a",
+    "alpha": 0.8,
+    "zorder": 1,
+	"colormap": "viridis"
+    },
+    "vert": {
+    "visible": false,
+    "s": 100,
+    "color": "#000a4b",
+    "alpha": 0.3,
+    "zorder": 2
+    },
+    "grad": {
+    "color":"#000a4b",
+    "alpha":0.5,
+    "width":0.04
+    },
+    "face": {
+    "visible": false,
+    "color":"#8aa678",
+    "alpha": 1.0,
+    "zorder": -1
+    },
+    "axis": {
+    "autoscale": true,
+    "color_bar_cmap":"viridis",
+    "color_bar_range":false, 
+    "color_bar_label":false, 
+    "color_bar_target":"face"}
+    }
     """
     draw_specs = sheet_spec()
     spec_updater(draw_specs, draw_specs_kw)
@@ -225,21 +239,28 @@ def sheet_view_GC_colorbar(sheet, coords=COORDS, ax1=None, ax2=None, **draw_spec
 
     # gc
     ax2 = fig.add_subplot(grid0[:, 9])
-    cmap = cm.get_cmap("viridis")
-    import matplotlib as mpl
-    # norm = mpl.colors.Normalize(vmin=np.min(draw_specs['face']['color']), vmax=np.max(draw_specs['face']['color']))
-    if draw_specs["face"]["color_bar_range"]==False:
-        norm = mpl.colors.Normalize(
-            vmin=np.min(sheet.face_df["col"]), vmax=np.max(sheet.face_df["col"])
-        )
+    cmap = cm.get_cmap(draw_specs["axis"]["color_bar_cmap"])
+
+    if draw_specs["axis"]["color_bar_range"] == False:
+        if draw_specs["axis"]["color_bar_target"] == "face":
+            norm = mpl.colors.Normalize(
+                vmin=np.min(sheet.face_df["col"]), vmax=np.max(sheet.face_df["col"])
+            )
+        else:
+            print(
+                "automatic color bar range only works for the face dataframe use draw_specs dictionary under axis, color_bar_range entry to specify a range"
+            )
     else:
         norm = mpl.colors.Normalize(
-            vmin=draw_specs["face"]["color_bar_range"][0], vmax=draw_specs["face"]["color_bar_range"][1]
[…]
