### 2021-03-03 #2973 @sophietheis
Hi,
I am working on demo notebook, and I am also starting to improve the manuscript according to @SergeDmi comments.
Modifications are on my remote, I will start a PR tomorrow.
