+++ paper.bib
@@ -1,8 +1,8 @@
-@article{Martin:2020,
+@article{Martin:2021,
 author = {Martin, Emmanuel and Theis, Sophie and Monier, Bruno and Gay, Guillaume and Suzanne, Magali},
-title = {{Mechanical control of morphogenesis robustness}},
-url = {https://www.biorxiv.org/content/10.1101/2020.01.06.896266v1).},
-doi = {10.1101/2020.01.06.896266}
+title = {{Arp2/3-dependent mechanical control of morphogenetic robustness in an inherently challenging environment}},
+url = {https://www.sciencedirect.com/science/article/pii/S1534580721000368?via%3Dihub.},
+doi = {10.1016/j.devcel.2021.01.005}
 }
 
 @article{Gracia:2019,
