### 2021-03-04 7bb2e5a
Add figure to show the independence between topology and mechanics
