### 2021-03-04 730c70d
Fix figure showing
