+++ paper.md
@@ -33,8 +33,8 @@ The `tyssue` Python library seeks to provide a unified interface to implement bi
 # Statement of Need
 <div align="justify">
 
-Tissue remodelling is a complex process integrating a large number of input such as gene expression pattern, cell adherent properties, cell mechanics. It can be difficult to manipulate specific aspects genetically. It can even be hard to simply capture, when the process takes only few minutes. Furthermore, morphogenesis is inherently a mechanical process. To execute complex morphogenetic movements, epithelia are driven by in-plane forces, like constriction of apical cell surface [@Heer:2017], and/or out-plane forces, such as the apico-basal cable in apoptotic cell [@Monier:2015, @Gracia:2019] or lateral tension [@Sherrard:2010, @Sui:2018]. Modeling those process help us understand how tissue acquires their shape, in complement of the experimental systems, and beyond their limitations. Several vertex models have been developed in the past few years to describe the physics of epithelia (for a review, see [@Alt:2017]), and common features can be identified. Several kind of model have already been published. The apical vertex model has been used several time to study topology change during morphogenetic movement in _Drosophila, Hydra and Xenopus_([@Staple:2010], [@Fahradifar:2007], [@Aegerter-Wilmsen:2010]). Associated with protein dynamics, it has been used to study the effect of protein position on tissue organisation in zebrafish retina ([@Salbreux:2012]). 
-3D vertex model have been used to study epithelium deformation due to normal development or to cancer development ([@Okuda:2014], [@Eritano:2020]).
+Tissue remodelling is a complex process integrating a large number of input such as gene expression pattern, cell adherent properties, cell mechanics. It can be difficult to manipulate specific aspects genetically. It can even be hard to simply capture, when the process takes only few minutes. Furthermore, morphogenesis is inherently a mechanical process. To execute complex morphogenetic movements, epithelia are driven by in-plane forces, like constriction of apical cell surface [@Heer:2017], and/or out-plane forces, such as the apico-basal cable in apoptotic cell [@Monier:2015, @Gracia:2019] or lateral tension [@Sherrard:2010, @Sui:2018]. Modeling those process help us understand how tissue acquires their shape, in complement of the experimental systems, and beyond their limitations. Several vertex models have been developed in the past few years to describe the physics of epithelia (for a review, see [@Alt:2017]), and common features can be identified. Several kind of model have already been published. The apical vertex model has been used several time to study topology change during morphogenetic movement in _Drosophila, Hydra and Xenopus_([@Staple:2010], [@Farhadifar:2007], [@Aegerter:2012]). Associated with protein dynamics, it has been used to study the effect of protein position on tissue organisation in zebrafish retina ([@Salbreux:2012]). 
+3D vertex model have been used to study epithelium deformation due to normal development or to cancer development ([@Okuda:2015], [@Eritano:2020]).
 Most of the time, models are developed for a specific biological question and are difficult to adapt to an other system, for severals reasons. With `tyssue` library, we propose models which are adaptable and scalable with the fied research and the biological question. Topology and mechanics are implement independantly to improve the versatility of models.
 
[…]
