+++ paper.bib
@@ -5,6 +5,17 @@ url = {https://www.sciencedirect.com/science/article/pii/S1534580721000368?via%3
 doi = {10.1016/j.devcel.2021.01.005}
 }
 
+@article{Eritano:2020,
+author = {Eritano, Anthony S. and Bromley, Claire L. and {Bolea Albero}, Antonio and Sch{\"{u}}tz, Lucas and Wen, Fu-Lai and Shibata, Tatsuo and Takeda, Michiko and Sami, Mustafa M. and Lemke, Steffen and Wang, Yu-Chiun},
+doi = {10.2139/ssrn.3378007},
+file = {:home/admin-suz/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Eritano et al. - 2020 - Tissue-Scale Mechanical Coupling Reduces Morphogenetic Noise to Ensure Precision During Epithelial Folding.pdf:pdf},
+journal = {Developmental cell},
+pages = {1--17},
+title = {{Tissue-Scale Mechanical Coupling Reduces Morphogenetic Noise to Ensure Precision During Epithelial Folding}},
+volume = {53},
+year = {2020}
+}
+
 @article{Gracia:2019,
 author = {Gracia, M{\'{e}}lanie and Theis, Sophie and Proag, Amsha and Gay, Guillaume and Benassayag, Corinne and Suzanne, Magali},
 doi = {10.1038/s41467-019-10720-0},
@@ -51,7 +62,6 @@ url = {http://dx.doi.org/10.1038/s41467-018-06497-3 http://www.nature.com/articl
 volume = {9},
 year = {2018}
 }
-
 
 @article{Heer:2017,
 author = {Heer, Natalie C and Miller, Pearson W. and Chanet, Soline and Stoop, Norbert and Dunkel, J{\"{o}}rn and Martin, Adam C},
@@ -151,6 +161,47 @@ volume = {12},
 year = {2012}
 }
 
+@article{Salbreux:2012,
+author = {Salbreux, Guillaume and Barthel, Linda K. and Raymond, Pamela A. and Lubensky, David K.},
+doi = {10.1371/journal.pcbi.1002618},
+issn = {1553-7358},
+journal = {PLoS Computational Biology},
+month = {aug},
+number = {8},
+pages = {e1002618},
+publisher = {Public Library of Science},
+title = {{Coupling Mechanical Deformations and Planar Cell Polarity to Create Regular Patterns in the Zebrafish Retina}},
+url = {http://dx.plos.org/10.1371/journal.pcbi.1002618},
+volume = {8},
+year = {2012}
+}
+
+@article{Aegerter:2012,
+author = {Aegerter-Wilmsen, Tinri and Heimlicher, Maria B. and Smith, Alister C. and de Reuille, Pierre Barbier and Smith, Richard S. and Aegerter, Christof M. and Konrad, Basler},
+doi = {10.1242/dev.082800},
+file = {:home/admin-suz/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Aegerter-Wilmsen et al. - 2012 - Integrating force-sensing and signaling pathways in a model for the regulation of wing imaginal disc si.pdf:pdf},
+journal = {Development},
+title = {{Integrating force-sensing and signaling pathways in a model for the regulation of wing imaginal disc size}},
+url = {http://dev.biologists.org/content/develop/139/17/3221.full.pdf?with-ds=yes},
+year = {2012}
+}
+
+@article{Staple:2010,
+doi = {10.1016/j.cell.2010.07.042},
+issn = {00928674},
+journal = {Cell},
+keywords = {Devbio},
+month = {sep},
[…]
