### 2021-03-11 be5fe21
update paper.md, mainly typos
