+++ paper.bib
@@ -145,6 +145,25 @@ volume = {15},
 year = {2015}
 }
 
+@article{Mirams:2013,
+author = {Mirams, Gary R. and Arthurs, Christopher J. and Bernabeu, Miguel O. and Bordas, Rafel and Cooper, Jonathan and Corrias, Alberto and Davit, Yohan and Dunn, Sara-Jane and Fletcher, Alexander G. and Harvey, Daniel G. and Marsh, Megan E. and Osborne, James M. and Pathmanathan, Pras and Pitt-Francis, Joe and Southern, James and Zemzemi, Nejib and Gavaghan, David J.},
+doi = {10.1371/journal.pcbi.1002970},
+editor = {Prlic, Andreas},
+issn = {1553-7358},
+journal = {PLoS Computational Biology},
+keywords = {Cardiac electrophysiology,Colorectal cancer,Computational biology,Heart,Open source software,Oxygen,Simulation and modeling,Source code},
+mendeley-groups = {Modeling},
+month = {mar},
+number = {3},
+pages = {e1002970},
+publisher = {Public Library of Science},
+title = {{Chaste: An Open Source C++ Library for Computational Physiology and Biology}},
+url = {https://dx.plos.org/10.1371/journal.pcbi.1002970},
+volume = {9},
+year = {2013}
+}
+
+
 @article{Okuda:2012,
 author = {Okuda, Satoru and Inoue, Yasuhiro and Eiraku, Mototsugu and Sasai, Yoshiki and Adachi, Taiji},
 doi = {10.1007/s10237-012-0430-7},
