### 2021-03-12 f2f0cc1
Update paper  (#225)
* Add legend and more software mention

* Add figure to show the independence between topology and mechanics

* Fix figure showing

* Fix biblio

* update paper.md, mainly typos

* Add mention to Chaste software

Co-authored-by: Guillaume Gay <guillaume@damcb.com>
