+++ paper.bib
@@ -1,8 +1,19 @@
-@article{Martin:2020,
+@article{Martin:2021,
 author = {Martin, Emmanuel and Theis, Sophie and Monier, Bruno and Gay, Guillaume and Suzanne, Magali},
-title = {{Mechanical control of morphogenesis robustness}},
-url = {https://www.biorxiv.org/content/10.1101/2020.01.06.896266v1).},
-doi = {10.1101/2020.01.06.896266}
+title = {{Arp2/3-dependent mechanical control of morphogenetic robustness in an inherently challenging environment}},
+url = {https://www.sciencedirect.com/science/article/pii/S1534580721000368?via%3Dihub.},
+doi = {10.1016/j.devcel.2021.01.005}
+}
+
+@article{Eritano:2020,
+author = {Eritano, Anthony S. and Bromley, Claire L. and {Bolea Albero}, Antonio and Sch{\"{u}}tz, Lucas and Wen, Fu-Lai and Shibata, Tatsuo and Takeda, Michiko and Sami, Mustafa M. and Lemke, Steffen and Wang, Yu-Chiun},
+doi = {10.2139/ssrn.3378007},
+file = {:home/admin-suz/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Eritano et al. - 2020 - Tissue-Scale Mechanical Coupling Reduces Morphogenetic Noise to Ensure Precision During Epithelial Folding.pdf:pdf},
+journal = {Developmental cell},
+pages = {1--17},
+title = {{Tissue-Scale Mechanical Coupling Reduces Morphogenetic Noise to Ensure Precision During Epithelial Folding}},
+volume = {53},
+year = {2020}
 }
 
 @article{Gracia:2019,
@@ -51,7 +62,6 @@ url = {http://dx.doi.org/10.1038/s41467-018-06497-3 http://www.nature.com/articl
 volume = {9},
 year = {2018}
 }
-
 
 @article{Heer:2017,
 author = {Heer, Natalie C and Miller, Pearson W. and Chanet, Soline and Stoop, Norbert and Dunkel, J{\"{o}}rn and Martin, Adam C},
@@ -135,6 +145,25 @@ volume = {15},
 year = {2015}
 }
 
+@article{Mirams:2013,
+author = {Mirams, Gary R. and Arthurs, Christopher J. and Bernabeu, Miguel O. and Bordas, Rafel and Cooper, Jonathan and Corrias, Alberto and Davit, Yohan and Dunn, Sara-Jane and Fletcher, Alexander G. and Harvey, Daniel G. and Marsh, Megan E. and Osborne, James M. and Pathmanathan, Pras and Pitt-Francis, Joe and Southern, James and Zemzemi, Nejib and Gavaghan, David J.},
+doi = {10.1371/journal.pcbi.1002970},
+editor = {Prlic, Andreas},
+issn = {1553-7358},
+journal = {PLoS Computational Biology},
+keywords = {Cardiac electrophysiology,Colorectal cancer,Computational biology,Heart,Open source software,Oxygen,Simulation and modeling,Source code},
+mendeley-groups = {Modeling},
+month = {mar},
+number = {3},
+pages = {e1002970},
+publisher = {Public Library of Science},
+title = {{Chaste: An Open Source C++ Library for Computational Physiology and Biology}},
+url = {https://dx.plos.org/10.1371/journal.pcbi.1002970},
+volume = {9},
+year = {2013}
+}
+
+
 @article{Okuda:2012,
 author = {Okuda, Satoru and Inoue, Yasuhiro and Eiraku, Mototsugu and Sasai, Yoshiki and Adachi, Taiji},
 doi = {10.1007/s10237-012-0430-7},
[…]
