### 2021-03-12 #2973 @sophietheis
Hi @SergeDmi

You can find in the [PR](https://github.com/DamCB/tyssue/pull/225) the modification made to our manuscript according to your comments.

1. Statement of need
_The statement of need does a good job of explaining the context, but lacks in describing the actual need for the software. A mention of existing software is also lacking. Instead, this paragraph gives a short description of Tyssue._

**We are agree that there is a lack of information about the need and the existing software. We have expanded this part with reference with several publish work :**

Several vertex models have been developed in the past few years to describe the physics of epithelia (for a review, see [@Alt:2017]), and common features can be identified. Several kinds of models have already been published. The apical vertex model has been used several times to study topology changes during morphogenetic movement in _Drosophila, Hydra and Xenopus_([@Staple:2010], [@Farhadifar:2007], [@Aegerter:2012]). Associated with protein dynamics, it has been used to study the effect of protein position on tissue organisation in zebrafish retina ([@Salbreux:2012]). 3D vertex model have been used to study epithelium deformation due to normal development or to cancer development ([@Okuda:2015], [@Eritano:2020]).
Most of the time, models are developed for a specific biological question and are difficult to adapt to an other system, for severals reasons. With the `tyssue` library, we propose models which are adaptable and scalable with the fied research and the biological question. Topology and mechanics are implement independantly to improve the versatility of models.

2. Software description
_I think the article fails to convey the organization of Tyssue, that "Separates structure, geometry and models" (better explained in the readme). As a consequence, it is not clear that the models (referenced in the paragraph mechanics) can be chosen at will between Euler, Gradient descent, Quasi-static, and possibly other and user defined ; the same applies to topology._

**We add a scheme into the manuscript to highlight the modularity of our library ([new scheme](https://github.com/DamCB/tyssue/blob/master/doc/illus/Figure2.png)).**
