### 2021-03-18 c0244a2
Fix `spec_updater`according to its docstring
