### 2021-03-19 f33d484
Simplify loop & add test
