+++ tests/draw/test_plt.py
@@ -15,7 +15,7 @@ from tyssue.generation import three_faces_sheet
 from tyssue import Sheet, config
 from tyssue.draw.plt_draw import quick_edge_draw, sheet_view
 from tyssue.draw.plt_draw import _face_color_from_sequence
-from tyssue.draw.plt_draw import create_gif
+from tyssue.draw.plt_draw import create_gif, plot_forceœ
 
 
 class TestsPlt:
@@ -51,7 +51,7 @@ class TestsPlt:
             0.0, 1.0, num=self.sheet.edge_df.shape[0]
         )[::-1]
 
-        self.draw_specs["edge"]["visible"] = True
+        self.draw_specs["edge"]["visible"] = True 
         self.draw_specs["edge"]["color"] = self.sheet.edge_df["rand"]  # [0, 0, 0, 1]
         self.draw_specs["edge"]["alpha"] = 1.0
         self.draw_specs["edge"]["color_range"] = 0, 3
@@ -155,3 +155,16 @@ def test_create_gif():
 
     os.remove("frames.gif")
     os.remove("interval.gif")
+
+
+def test_plot_force():
+    geom = SheetGeometry
+    model = PlanarModel
+    sheet = Sheet("3", *three_faces_sheet())
+    geom.update_all(sheet)
+    fig, ax = plot_force(sheet,
+                         geom,
+                         model,
+                         list('xy'),
+                         0.5,
+                         **{'extract': {'z_boundary': (-10, 10)}})
