### 2021-03-19 c52ad0e
Improve `plot_force` and fix `spec_updater` (#227)
* Fix `spec_updater`according to its docstring

* Update `plot_force` function to allows crop tyssue

* Add test

* Simplify loop & add test

* Fix typo error

* There is more than one force...

* add test
