+++ doc/index.rst
@@ -1,34 +1,0 @@
-.. Tyssue documentation master file, created by
-   sphinx-quickstart on Mon Aug 22 16:57:50 2016.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to Tyssue's documentation!
-==================================
-
-
-The ``tyssue`` library seeks to provide a unified interface to implement
-bio-mechanical models of living tissues, with a focus on **epithelium**
-modeling. It's main focus is on **vertex** based models.
-
-.. toctree::
-   :maxdepth: 2
-
-   Basic creation and index manipulation <notebooks/Basic_Usage>
-   Visualisation <notebooks/Visualization>
-   Geometries <notebooks/Geometries_showcase>
-   Quasistatic Models <notebooks/Quasistatic_models>
-   Viscous models <notebooks/SimpleTimeDependent>
-   Cell Division <notebooks/Cell_Division>
-   Type 1 transition <notebooks/Type_1_transition>
-   Apoptosis <notebooks/Apoptosis>
-   API
-
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
