+++ doc/index.md
@@ -1,0 +1,6 @@
+# Welcome to Tyssue's documentation!
+
+
+The `tyssue` library seeks to provide a unified interface to implement
+bio-mechanical models of living tissues.
+It's main focus is on **vertex based epithelium models**.
