### 2021-03-21 #2973 @fcooper8472
My apologies for the late review: I had somehow managed to turn off notifications for this review, so had forgotten about it until Pierre emailed me.  Sorry!

---
I have now completed my review.

The software clearly passes the threshold scholarly effort.  It appears to be well engineered, easy to install, and very thoroughly documented including with extensive notebooks.

The paper itself is well written and explains the purpose and need of the software.

I recommend it for publication in JOSS.

---

The only minor comments I have are:

- Figure 1 (in particular) and Figure 2 (to a lesser extent) are pixelated in the paper. Is it possible to produce a higher-resolution version?
- The README contains the following, which doesn't seem to make chronological sense:
  ```
  Since version 0.3, this project is distributed under the terms of the General Public Licence.
  Versions 2.4 and earlier were distributed under the Mozilla Public Licence.
  ```
- There is a more recent publication for the Chaste software (https://doi.org/10.21105/joss.01848) which is now preferred for general citations, although I should state my conflict of interest as the first author on that paper
