### 2021-03-22 #2973 @glyg
Thank you for the review @fcooper8472 , I'll fix the version problem, and we'll output higher resolution bitmaps for the figures, and fix the citation for Chaste

We are finishing a revamping of the documentation, it should be done by the end of the week, to fully account for @SergeDmi remarks, I'll comment here when it's done.
