### 2021-03-22 9465d06
Update bibliography and figures (#228)
* Update bibliography

* Update fig + less pixelated images
