+++ paper.bib
@@ -16,6 +16,23 @@ volume = {53},
 year = {2020}
 }
 
+@article{Cooper:2020,
+author = {Cooper, Fergus and Baker, Ruth and Bernabeu, Miguel and Bordas, Rafel and Bowler, Louise and Bueno-Orovio, Alfonso and Byrne, Helen and Carapella, Valentina and Cardone-Noott, Louie and Cooper, Jonathan and Dutta, Sara and Evans, Benjamin and Fletcher, Alexander and Grogan, James and Guo, Wenxian and Harvey, Daniel and Hendrix, Maurice and Kay, David and Kursawe, Jochen and Maini, Philip and McMillan, Beth and Mirams, Gary and Osborne, James and Pathmanathan, Pras and Pitt-Francis, Joe and Robinson, Martin and Rodriguez, Blanca and Spiteri, Raymond and Gavaghan, David},
+doi = {10.21105/joss.01848},
+issn = {2475-9066},
+journal = {Journal of Open Source Software},
+mendeley-groups = {Modeling},
+month = {mar},
+number = {47},
+pages = {1848},
+publisher = {The Open Journal},
+title = {{Chaste: Cancer, Heart and Soft Tissue Environment}},
+url = {https://joss.theoj.org/papers/10.21105/joss.01848},
+volume = {5},
+year = {2020}
+}
+
+
 @article{Gracia:2019,
 author = {Gracia, M{\'{e}}lanie and Theis, Sophie and Proag, Amsha and Gay, Guillaume and Benassayag, Corinne and Suzanne, Magali},
 doi = {10.1038/s41467-019-10720-0},
@@ -144,25 +161,6 @@ url = {http://link.springer.com/10.1007/s10237-015-0724-7},
 volume = {15},
 year = {2015}
 }
-
-@article{Mirams:2013,
-author = {Mirams, Gary R. and Arthurs, Christopher J. and Bernabeu, Miguel O. and Bordas, Rafel and Cooper, Jonathan and Corrias, Alberto and Davit, Yohan and Dunn, Sara-Jane and Fletcher, Alexander G. and Harvey, Daniel G. and Marsh, Megan E. and Osborne, James M. and Pathmanathan, Pras and Pitt-Francis, Joe and Southern, James and Zemzemi, Nejib and Gavaghan, David J.},
-doi = {10.1371/journal.pcbi.1002970},
-editor = {Prlic, Andreas},
-issn = {1553-7358},
-journal = {PLoS Computational Biology},
-keywords = {Cardiac electrophysiology,Colorectal cancer,Computational biology,Heart,Open source software,Oxygen,Simulation and modeling,Source code},
-mendeley-groups = {Modeling},
-month = {mar},
-number = {3},
-pages = {e1002970},
-publisher = {Public Library of Science},
-title = {{Chaste: An Open Source C++ Library for Computational Physiology and Biology}},
-url = {https://dx.plos.org/10.1371/journal.pcbi.1002970},
-volume = {9},
-year = {2013}
-}
-
 
 @article{Okuda:2012,
 author = {Okuda, Satoru and Inoue, Yasuhiro and Eiraku, Mototsugu and Sasai, Yoshiki and Adachi, Taiji},
