### 2021-04-01 #2973 @glyg
Hi @pdebuyl, sorry for the delay (all my fault), I'm polishing the doc tomorrow and deploying it (we switched from readthedocs to jupyter books)
