+++ doc/_toc.yml
@@ -4,12 +4,11 @@
 - file: notebooks/02-Visualization.ipynb
 - file: notebooks/03-Dynamics.ipynb
 - file: notebooks/04-Solvers.ipynb
-- file: notebooks/05-EventManager.ipynb
-- file: notebooks/06-Type_1_transition.ipynb
-- file: notebooks/08-Rosettes.ipynb
-- file: notebooks/12-FarhadifarModel.ipynb
+- file: notebooks/05-Rearangments.ipynb
+- file: notebooks/06-Cell_Division.ipynb
+- file: notebooks/07-EventManager.ipynb
 - file: notebooks/A-ReversibleNetworkReconnection.ipynb
-- file: notebooks/B-Cell_Division.ipynb
-- file: notebooks/C-Apoptosis.ipynb
-- file: notebooks/D-2DMigration.ipynb
+- file: notebooks/B-Apoptosis.ipynb
+- file: notebooks/C-2DMigration.ipynb
+- file: notebooks/D-FarhadifarModel.ipynb
 - file: API
