+++ .travis.yml
@@ -1,4 +1,4 @@
-dist: xenial
+dist: focal
 
 before_install:
   - sudo apt-get update
