+++ paper.bib
@@ -1,292 +1,0 @@
-@article{Martin:2021,
-author = {Martin, Emmanuel and Theis, Sophie and Monier, Bruno and Gay, Guillaume and Suzanne, Magali},
-title = {{Arp2/3-dependent mechanical control of morphogenetic robustness in an inherently challenging environment}},
-url = {https://www.sciencedirect.com/science/article/pii/S1534580721000368?via%3Dihub.},
-doi = {10.1016/j.devcel.2021.01.005}
-}
-
-@article{Eritano:2020,
-author = {Eritano, Anthony S. and Bromley, Claire L. and {Bolea Albero}, Antonio and Sch{\"{u}}tz, Lucas and Wen, Fu-Lai and Shibata, Tatsuo and Takeda, Michiko and Sami, Mustafa M. and Lemke, Steffen and Wang, Yu-Chiun},
-doi = {10.2139/ssrn.3378007},
-file = {:home/admin-suz/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Eritano et al. - 2020 - Tissue-Scale Mechanical Coupling Reduces Morphogenetic Noise to Ensure Precision During Epithelial Folding.pdf:pdf},
-journal = {Developmental cell},
-pages = {1--17},
-title = {{Tissue-Scale Mechanical Coupling Reduces Morphogenetic Noise to Ensure Precision During Epithelial Folding}},
-volume = {53},
-year = {2020}
-}
-
-@article{Cooper:2020,
-author = {Cooper, Fergus and Baker, Ruth and Bernabeu, Miguel and Bordas, Rafel and Bowler, Louise and Bueno-Orovio, Alfonso and Byrne, Helen and Carapella, Valentina and Cardone-Noott, Louie and Cooper, Jonathan and Dutta, Sara and Evans, Benjamin and Fletcher, Alexander and Grogan, James and Guo, Wenxian and Harvey, Daniel and Hendrix, Maurice and Kay, David and Kursawe, Jochen and Maini, Philip and McMillan, Beth and Mirams, Gary and Osborne, James and Pathmanathan, Pras and Pitt-Francis, Joe and Robinson, Martin and Rodriguez, Blanca and Spiteri, Raymond and Gavaghan, David},
-doi = {10.21105/joss.01848},
-issn = {2475-9066},
-journal = {Journal of Open Source Software},
-mendeley-groups = {Modeling},
-month = {mar},
-number = {47},
-pages = {1848},
-publisher = {The Open Journal},
-title = {{Chaste: Cancer, Heart and Soft Tissue Environment}},
-url = {https://joss.theoj.org/papers/10.21105/joss.01848},
-volume = {5},
-year = {2020}
-}
-
-
-@article{Gracia:2019,
-author = {Gracia, M{\'{e}}lanie and Theis, Sophie and Proag, Amsha and Gay, Guillaume and Benassayag, Corinne and Suzanne, Magali},
-doi = {10.1038/s41467-019-10720-0},
-issn = {2041-1723},
-journal = {Nature Communications},
-keywords = {Epithelial,Mesoderm,Myosin,mesenchymal transition},
-month = {dec},
-number = {1},
-pages = {2951},
-publisher = {Nature Publishing Group},
-title = {{Mechanical impact of epithelial−mesenchymal transition on epithelial morphogenesis in Drosophila}},
-url = {http://www.nature.com/articles/s41467-019-10720-0},
-volume = {10},
-year = {2019}
-}
-
-@article{Finegan:2019,
[…]
