+++ paper.md
@@ -1,98 +1,0 @@
----
-title: 'Tyssue: an epithelium simulation library'
-tags:
-  - Python
-  - Developmental biology
-  - Epithelium
-authors:
-  - name: Sophie Theis
-    orcid: 0000-0003-4206-8153
-    affiliation: "1, 2"
-  - name: Magali Suzanne
-    orcid: 0000-0003-0801-3706
-    affiliation: "2"
-  - name: Guillaume Gay
-    orcid: 0000-0002-3957-2474
-    affiliation: "1, 3"
-affiliations:
- - name: Morphogénie Logiciels, 32110 St Martin d’Armagnac, France.
-   index: 1
- - name: LBCMCP, Centre de Biologie Intégrative (CBI), Université de Toulouse, CNRS, UPS, Toulouse 31062, France.
-   index: 2
- - name: Turing Center for Living Systems, Marseille, France
-   index: 3
-date: 2 december 2020
-bibliography: paper.bib
-
----
-# Summary
-<div align="justify">
-The `tyssue` Python library seeks to provide a unified interface to implement bio-mechanical models of living tissues. It's main focus is on vertex based epithelium models. `tyssue` allows to model the mechanical behavior of 2D, apical 3D or full 3D epihelia based on the numerical resolution of the equations of motion for the mesh vertices. Biological process are modeled through changes in the topological and dynamical properties of the mesh. `tyssue` is a modular library. Starting with the same tissue geometry, the choice of constraints, energy potential terms and parameters increases the possibility to answer different biological questions and easily explore mechanical hypotheses.
-</div>
-
-# Statement of Need
-<div align="justify">
-
-Tissue remodelling is a complex process integrating a large number of input such as gene expression pattern, cell adherent properties, cell mechanics. It can be difficult to manipulate specific aspects genetically. It can even be hard to simply capture, when the process takes only few minutes. Furthermore, morphogenesis is inherently a mechanical process. To execute complex morphogenetic movements, epithelia are driven by in-plane forces, like constriction of apical cell surface [@Heer:2017], and/or out-of-plane forces, such as the apico-basal cable in apoptotic cell [@Monier:2015, @Gracia:2019] or lateral tension [@Sherrard:2010, @Sui:2018]. Modeling these processes help us understand how tissue acquire their shape, in complement of the experimental systems, and beyond their limitations. Several vertex models have been developed in the past few years to describe the physics of epithelia (for a review, see [@Alt:2017]), and common features can be identified. Several kinds of models have already been published. The apical vertex model has been used several times to study topology changes during morphogenetic movement in _Drosophila, Hydra and Xenopus_([@Staple:2010], [@Farhadifar:2007], [@Aegerter:2012]). Associated with protein dynamics, it has been used to study the effect of protein position on tissue organisation in zebrafish retina ([@Salbreux:2012]). 3D vertex model have been used to study epithelium deformation due to normal development or to cancer development ([@Okuda:2015], [@Eritano:2020]). Most of the time, models are developed for a specific biological question and are difficult to adapt to an other system, for severals reasons. However, there is some exception like Chaste [@Cooper:2020], which propose an open source C++ library to model cell populations or how specific events arise at the system level. With the `tyssue` library, we propose models which are adaptable and scalable with the fied research and the biological question. Topology and mechanics are implement independantly to improve the versatility of models.
-
[…]
