### 2021-04-02 #2973 @glyg
Hi @pdebuyl @fcooper8472 and @SergeDmi - we updated the doc, now at [tyssue.io](http://tyssue.io), hopefully it is more detailed and progressive.

We also fixed the citation od Chaste and the README issue
