### 2021-04-14 #2973 @kyleniemeyer
Hi there, I'm the AEIC on duty this week, just looking to help submissions get past the finish line. It looks like @glyg and @sophietheis made some edits to the article in response to @SergeDmi's comments, so I'm thinking this may be nearly ready to accept?

I did notice some issues with references, that you can address:
- the "Cell Flow Reorients..." citation seems to be missing authors
- The Honda (1983) citation is missing the journal
- the Martin et al. citation is missing year and journal
