### 2021-04-15 b4d55e9
Fixes bib issues
see https://github.com/openjournals/joss-reviews/issues/2973#issuecomment-819744476
