+++ paper.bib
@@ -1,6 +1,9 @@
 @article{Martin:2021,
-author = {Martin, Emmanuel and Theis, Sophie and Monier, Bruno and Gay, Guillaume and Suzanne, Magali},
-title = {{Arp2/3-dependent mechanical control of morphogenetic robustness in an inherently challenging environment}},
+title = {Arp2/3-Dependent Mechanical Control of Morphogenetic Robustness in an Inherently Challenging Environment},
+author = {Martin, Emmanuel and Theis, Sophie and Gay, Guillaume and Monier, Bruno and Rouvière, Christian and Suzanne, Magali},
+date = {2021-02-02},
+journaltitle = {Developmental Cell},
+issn = {1534-5807},
 url = {https://www.sciencedirect.com/science/article/pii/S1534580721000368?via%3Dihub.},
 doi = {10.1016/j.devcel.2021.01.005}
 }
@@ -204,19 +207,16 @@ year = {2012}
 }
 
 @article{Staple:2010,
-doi = {10.1016/j.cell.2010.07.042},
-issn = {00928674},
-journal = {Cell},
-keywords = {Devbio},
-month = {sep},
-number = {5},
-pages = {773--786},
-pmid = {20813263},
-publisher = {Cell Press},
 title = {{Cell Flow Reorients the Axis of Planar Polarity in the Wing Epithelium of Drosophila}},
-url = {https://www-sciencedirect-com.insb.bib.cnrs.fr/science/article/pii/S0092867410008901},
+author = {Aigouy, Benoît and Farhadifar, Reza and Staple, Douglas B. and Sagner, Andreas and Röper, Jens-Christian and Jülicher, Frank and Eaton, Suzanne},
+date = {2010-09-03},
+journal = {Cell},
 volume = {142},
-year = {2010}
+pages = {773--786},
+issn = {0092-8674},
+doi = {10.1016/j.cell.2010.07.042},
+url = {https://www.sciencedirect.com/science/article/pii/S0092867410008901},
+publisher = {Cell Press},
 }
 
 @article{Sherrard:2010,
@@ -267,7 +267,7 @@ year = {2002}
 @article{Honda:1983,
 title = "Geometrical Models for Cells in Tissues",
 editor = "G.H. Bourne and J.F. Danielli and K.W. Jeon",
-series = "International Review of Cytology",
+journal = "International Review of Cytology",
 publisher = "Academic Press",
 volume = "81",
 pages = "191 - 248",
