### 2021-04-15 7ee590f
Added the new functions for the pull request
