### 2021-04-15 c30e2ad
removed commented out portion in plt_draw.py
