### 2021-04-21 1a5049b
Fix color_bar option on or off plus error message when no color_bar_range is specified. Makes the script independent of our newly defined column in the face_df
