+++ tests/generation/test_generation.py
@@ -28,13 +28,19 @@ def test_from_3d_voronoi():
     datasets = from_3d_voronoi(Voronoi(grid))
     assert datasets["vert"].shape[0] == 139
     assert datasets["edge"].shape[0] == 1272
-    assert datasets["face"].shape[0] == 141
+    assert datasets["face"].shape[0] == 282
     assert datasets["cell"].shape[0] == 70
     bulk = Epithelium("bulk", datasets, config.geometry.bulk_spec())
     bulk.reset_index()
     bulk.reset_topo()
     BulkGeometry.update_all(bulk)
     bulk.sanitize()
+
+    # GH 137
+    assert (
+        bulk.edge_df.groupby("face").apply(lambda df: df["cell"].unique().size).max()
+        == 1
+    )
     assert bulk.validate()
 
 
@@ -88,7 +94,7 @@ def test_hexagrid3d_noise():
     datasets = from_3d_voronoi(Voronoi(grid))
     assert datasets["vert"].shape[0] == 318
     assert datasets["edge"].shape[0] == 3300
-    assert datasets["face"].shape[0] == 335
+    assert datasets["face"].shape[0] == 670
     assert datasets["cell"].shape[0] == 72
 
 
