### 2021-05-12 85a5fe0
cleans dtype, fixing deprecation warning from numpy
