### 2021-05-12 842887e
cleans dtype, fixing deprecation warning from numpy
