### 2021-05-12 ef8e015
corrects upcast bug, fixes #236
