### 2021-05-12 ce3ec3d
use relative pos in face_projected_pos, fixes #235
