+++ tyssue/geometry/sheet_geometry.py
@@ -6,8 +6,7 @@ from .utils import rotation_matrix, rotation_matrices
 
 
 class SheetGeometry(PlanarGeometry):
-    """Geometry definitions for 2D sheets in 3D
-    """
+    """Geometry definitions for 2D sheets in 3D"""
 
     @classmethod
     def update_all(cls, sheet):
@@ -62,8 +61,7 @@ class SheetGeometry(PlanarGeometry):
 
         """
         sheet.edge_df["sub_vol"] = (
-            sheet.upcast_srce(
-                sheet.vert_df["height"]) * sheet.edge_df["sub_area"]
+            sheet.upcast_srce(sheet.vert_df["height"]) * sheet.edge_df["sub_area"]
         )
         sheet.face_df["vol"] = sheet.sum_face(sheet.edge_df["sub_vol"])
 
@@ -100,14 +98,12 @@ class SheetGeometry(PlanarGeometry):
             sheet.vert_df["rho"] = sheet.vert_df[w]
 
         elif geometry == "spherical":
-            sheet.vert_df["rho"] = np.linalg.norm(
-                sheet.vert_df[sheet.coords], axis=1)
+            sheet.vert_df["rho"] = np.linalg.norm(sheet.vert_df[sheet.coords], axis=1)
         elif geometry == "rod":
 
             a, b = sheet.settings["ab"]
             w0 = b - a
-            sheet.vert_df["rho"] = np.linalg.norm(
-                sheet.vert_df[[u, v]], axis=1)
+            sheet.vert_df["rho"] = np.linalg.norm(sheet.vert_df[[u, v]], axis=1)
             sheet.vert_df["left_tip"] = sheet.vert_df[w] < -w0
             sheet.vert_df["right_tip"] = sheet.vert_df[w] > w0
             l_mask = sheet.vert_df[sheet.vert_df["left_tip"] == 1].index
@@ -123,8 +119,7 @@ class SheetGeometry(PlanarGeometry):
         elif sheet.settings["geometry"] == "surfacic":
             sheet.vert_df["rho"] = 1.0
 
-        sheet.vert_df["height"] = sheet.vert_df[
-            "rho"] - sheet.vert_df["basal_shift"]
+        sheet.vert_df["height"] = sheet.vert_df["rho"] - sheet.vert_df["basal_shift"]
 
         edge_height = sheet.upcast_srce(sheet.vert_df[["height", "rho"]])
         edge_height.set_index(sheet.edge_df["face"], append=True, inplace=True)
@@ -166,8 +161,7 @@ class SheetGeometry(PlanarGeometry):
           The rotation matrix
 
         """
-        normal = sheet.edge_df[sheet.edge_df[
-            "face"] == face][sheet.ncoords].mean()
+        normal = sheet.edge_df[sheet.edge_df["face"] == face][sheet.ncoords].mean()
         normal = normal / np.linalg.norm(normal)
         n_xy = np.linalg.norm(normal[["nx", "ny"]])
         theta = -np.arctan2(n_xy, normal.nz)
@@ -199,22 +193,21 @@ class SheetGeometry(PlanarGeometry):
            The rotated, relative positions of the face's vertices
         """
 
-        face_orbit = sheet.edge_df[sheet.edge_df["face"] == face]["srce"]
-        rel_pos = (
-            sheet.vert_df.loc[face_orbit.to_numpy(), sheet.coords].to_numpy()
-            - sheet.face_df.loc[face, sheet.coords].to_numpy()
-        )
+        rel_pos = sheet.edge_df.query(f"face == {face}")[["srce", "rx", "ry", "rz"]]
+        rel_pos = rel_pos.set_index("srce")
+        rel_pos.index.name = "vert"
[…]
