### 2021-05-21 #2973 @glyg
Some links to notebooks were broken as we changed the organization of the doc, we're fixing them now
