### 2021-05-21 1cc5ba3
FIx link to demo notebooks and documentation (#241)
