### 2021-05-21 #2973 @pdebuyl
Hi, @glyg @sophietheis I'll proceed to the next step https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance

I found the following grammatical and spelling errors on the first page. Can you fix them and run a spell checker on the article?

- "It's main focus" -> "Its main focus"
- "epihelia": is it "epithelia" ? (added "t").
- First sentence in statement of need is a list without end. The last entry should start with a "and" or an "or", or end with etc.
- "helps us understand" -> "helps us to understand"
- "how tissue" -> "how tissues"
- "to an other system" -> "to another system" or "to other systems"
- "the fiel research" -> "the field of research"
- "are implement" -> "are implemented"
- "independantly" -> "independently"

Also, you should combine citations with a ";" inside square brackets. See the [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax).

You mention "the scientific Python tools" in the acknowledgments. Some of the core packages, among which NumPy, have citation guidelines: https://scipy.org/citing.html#numpy

Can you check the version number of tyssue associated with the submission, archive the code on zenodo (with correct metadata for the title and author list), and give an update here with the DOI?
