### 2021-05-21 #2973 @glyg
Thanks @pdebuyl we'll go through all that. I'll make a release of the code & check the zenodo metadata et al.
@sophietheis looks into the spelling and citations syntax.

We'll update you here (not today for me, I don't have enough time ^^)
