### 2021-05-27 #2973 @glyg
Hi @pdebuyl I just made a release of tyssue, tagged [0.9.0](https://github.com/DamCB/tyssue/releases/tag/0.9.0)  with the DOI [10.5281/zenodo.4817609](https://doi.org/10.5281/zenodo.4817609) and @sophietheis addressed the spelling  citation issues in [issue 242](https://github.com/DamCB/tyssue/pull/242).
