### 2021-06-03 #2973 @pdebuyl
@glyg can you update the metadata of the zenodo archive (you can "edit" it via the "upload" page) to match those of the paper (title, etc)? I see that the author listing does not match, can you address that?

Depending on the situation, some names can be listed as "contributors" and not "authors" depending on their contribution. "gitter-badger" should be removed. Check our authorship guidelines: https://joss.readthedocs.io/en/latest/submitting.html#authorship
