### 2021-06-03 #2973 @pdebuyl
Thank you for the changes. Can you remove html formatting in the md file?

There are still a few minor spelling errors, I'll try to go over the paper carefully in the coming days.
