### 2021-06-03 7f88891
Update paper.md
remove html formatting tags
