### 2021-06-03 4d8f862
Update paper.md
remove html formatting tags
