### 2021-06-03 #2973 @glyg
I removed the tags:
https://github.com/DamCB/tyssue/commit/4d8f8624b5d94649f0cd3e0be3a83def3d9f711f

I'll try to read it again also.
Thanks
