+++ paper.bib
@@ -35,7 +35,7 @@ volume = {5},
 year = {2020}
 }
 
-@ARTICLE{2020SciPy-NMeth,
+@ARTICLE{Virtanen:2020,
   author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
             Haberland, Matt and Reddy, Tyler and Cournapeau, David and
             Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
@@ -56,10 +56,10 @@ year = {2020}
   volume  = {17},
   pages   = {261--272},
   adsurl  = {https://rdcu.be/b08Wh},
-  doi     = {10.1038/s41592-019-0686-2},
+  doi     = {10.1038/s41592-019-0686-2}
 }
 
-@ARTICLE{2020NumPy-Array,
+@ARTICLE{Harris:2020,
   author  = {Harris, Charles R. and Millman, K. Jarrod and
             van der Walt, Stéfan J and Gommers, Ralf and
             Virtanen, Pauli and Cournapeau, David and
@@ -278,7 +278,7 @@ volume = {20},
 year = {2010}
 }
 
-@inproceedings{mckinney-proc-scipy-2010,
+@inproceedings{mckinney:2010,
 author = {McKinney, Wes},
 booktitle = {{P}roceedings of the 9th {P}ython in {S}cience {C}onference},
 doi = {10.25080/Majora-92bf1922-00a},
