+++ paper.md
@@ -17,7 +17,7 @@ authors:
 affiliations:
  - name: Morphogénie Logiciels, 32110 St Martin d’Armagnac, France.
    index: 1
- - name: MCD, Centre de Biologie Intégrative (CBI), Université de Toulouse, CNRS, UPS, Toulouse 31062, France.
+ - name: LBCMCP, Centre de Biologie Intégrative (CBI), Université de Toulouse, CNRS, UPS, Toulouse 31062, France.
    index: 2
  - name: Turing Center for Living Systems, Marseille, France
    index: 3
@@ -26,18 +26,21 @@ bibliography: paper.bib
 
 ---
 # Summary
-
+<div align="justify">
 The `tyssue` Python library seeks to provide a unified interface to implement bio-mechanical models of living tissues. Its main focus is on vertex based epithelium models. `tyssue` allows to model the mechanical behavior of 2D, apical 3D or full 3D epithelia based on the numerical resolution of the equations of motion for the mesh vertices. Biological processes are modeled through changes in the topological and dynamical properties of the mesh. `tyssue` is a modular library. Starting with the same tissue geometry, the choice of constraints, energy potential terms and parameters increases the possibility to answer different biological questions and easily explore mechanical hypotheses.
+</div>
 
 # Statement of Need
+<div align="justify">
 
-Tissue remodeling is a complex process integrating a large number of inputs such as gene expression patterns, cell adherent properties or cell mechanics. It can be difficult to manipulate specific aspects genetically. It can even be hard to simply capture, when the process takes only few minutes. Furthermore, morphogenesis is inherently a mechanical process. To execute complex morphogenetic movements, epithelia are driven by in-plane forces, like constriction of apical cell surface [@Heer:2017], and/or out-of-plane forces, such as lateral tension [@Sherrard:2010; @Sui:2018] or the apico-basal cable in apoptotic cells or cells undergoing epithelial to mesenchymal transition [@Monier:2015; @Gracia:2019]. Modeling these processes helps us understand how tissues acquire their shape, in complement of the experimental systems, and beyond their limitations. Several vertex models have been developed in the past few years to describe the physics of epithelia (for a review, see [@Alt:2017]), and common features can be identified. Several kinds of models have already been published. The apical vertex model has been used several times to study topology changes during morphogenetic movement in _Drosophila_, _Hydra_ and _Xenopus_[@Staple:2010; @Farhadifar:2007; @Aegerter:2012]. Associated with protein dynamics, it has been used to study the effect of protein position on tissue organisation in the zebrafish retina [@Salbreux:2012]. 3D vertex model have been used to study epithelium deformation due to normal or cancerous development [@Okuda:2015; @Eritano:2020]. Most of the time, models are developed for a specific biological question and are difficult to adapt to other systems, for several reasons, such as a lack of modularity. However, there are some exceptions like Chaste [@Cooper:2020], which proposes an open source C++ library to model cell populations or how specific events arise at the system level. With the `tyssue` library, we propose models which are adaptable and scalable with the field of research and the biological question. Topology and mechanics are implemented independently to improve the versatility of models.
[…]
