### 2021-06-03 #2973 @sophietheis
> Fixed parentheses & some typos
>
> @whedon generate pdf

Also fixed missing references.
