### 2021-06-07 7c3d43f
Remove html tags, fix citation syntax, minor edits
The html tags should not be included in the markdown source.

Some citations were enclosed in an extra set of parentheses which I
removed.

I added the section "features of tyssue" to close the "statement of
need" paragraph.

Minor grammatical fix "to how to" -> "on how to"
