### 2021-06-07 #2973 @pdebuyl
Hi @glyg and @sophietheis . I opened PR https://github.com/DamCB/tyssue/pull/244 with the remaining syntax issues. I added a section title for the "not statement of need" part of the paper and fix a minor grammatical error.

Can you have a look at it and let me know if it is suitable for you? A final request: in the acknowledgments, you include one of the co-authors. Can you change it to "We wish to thank the team of Magali Suzanne (...)"? The current situation is a bit weird.

Apart from these issues, I am good for approving the paper.
