### 2021-06-07 #2973 @glyg
Thanks a lot @pdebuyl - I thought I did remove the tags [here](https://github.com/DamCB/tyssue/commit/4d8f8624b5d94649f0cd3e0be3a83def3d9f711f#diff-0403ef06adf405f7b310b4518bd6a3559854f54c61676f676ce9cbfee7172ab6) but I guess there was some merge problem, sorry you had to do it again.

>  The current situation is a bit weird.

Ha yes good point, I saw it and was not sure how to get around it. I rephrased it now.
